+++ frontend/components/app.vue
@@ -1,0 +1,44 @@
+<template>
+    <div>
+        <div class="sixteen wide column">
+            <div class="ui secondary pointing menu">
+                <a class="item" v-link="{ name: 'cards', activeClass: 'active' }">Cards</a>
+                <a class="item" v-link="{ name: 'help', activeClass: 'active' }">Help</a>
+            </div>
+        </div>
+
+        <router-view></router-view>
+
+        <div class="sixteen wide right aligned column">
+            <div class="ui divider"></div>
+            <small class="secondary">this is <a href="https://github.com/pklink/machdas">machdas 0.6.2</a></small>
+        </div>
+    </div>
+</template>
+
+<script type="text/babel">
+    import key from 'keymaster'
+
+    export default {
+
+        ready() {
+            // create new task
+            key('n', () => {
+                this.$broadcast('tasks.new')
+                return false
+            })
+
+            // create new card
+            key('c', () => {
+                this.$broadcast('cards.new')
+                return false
+            })
+
+            // go to help
+            key('h', () => {
+                this.$route.router.go({ name: 'help' })
+            })
+        }
+
+    }
+</script>