+++ frontend/components/help.js
@@ -1,7 +1,0 @@
-import Vue from 'vue'
-
-export default Vue.extend({
-
-    template: require('./views/help.html')
-
-})